requirements.txt
python-dotenv==1.0.0
supabase==2.3.0
Pillow==10.2.0
requests==2.31.0
aiohttp==3.9.3
asyncio==3.4.3 